# backend/src/api/openapi-generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vibecode AI Planner API",
    "version": "1.0.0",
    "description": "API documentation for Vibecode AI Planner - project and ticket management with agent orchestration"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Agent API key for agent authentication"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/providers/{id}/agents": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get agents for a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents"
          }
        }
      }
    },
...
    "/memory/agent/{agentId}": {
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "Get memories for an agent",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agent memories"
          }
        }
      }
    },
...
    "/agents/create": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create a new agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created with API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "api_key": {
                      "type": "string"
                    },
                    "generatedApiKey": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Creation failed"
          }
        }
      }
    },
...
    "/agents/revoke/{agentId}": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Revoke agent API key",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked"
          },
          "400": {
            "description": "Revoke failed"
          }
        }
      }
    },
    "/agents/{agentId}": {
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Delete agent",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent deleted"
          },
          "400": {
            "description": "Deletion failed"
          }
        }
      }
    },
    "/agents/{agentId}/history": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get agent activity history",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent activity history"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/agents/{agentId}/key": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Get agent API key info (preview only)",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent key info"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/agents/:agentId/rotate-key": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Rotate agent API key",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key rotated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agentId": {
                      "type": "integer"
                    },
                    "agentName": {
                      "type": "string"
                    },
                    "newApiKey": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    }
  },
...
}